Upgrade a
- move `helpers` to `sspy`
- update/improve help
- add option for existing exports in current folder
- update readme
- add option for webroot
- move env-loading after changing working dir in __name__
- extend help display

+++ sspy/files.py
@@ -4,11 +4,11 @@ from pathlib import Path
 
 
 class Files:
-    def delete_sources(self, basepath):
+    def delete_sources(self, basepath, db, assets):
         try:
-            if os.path.isfile(os.path.join(basepath, 'assets.tar.gz')):
+            if not assets and os.path.isfile(os.path.join(basepath, 'assets.tar.gz')):
                 Path(os.path.join(basepath, 'assets.tar.gz')).unlink()
-            if os.path.isfile(os.path.join(basepath, 'database.sql.gz')):
+            if not db and os.path.isfile(os.path.join(basepath, 'database.sql.gz')):
                 Path(os.path.join(basepath, 'database.sql.gz')).unlink()
             return True
         except PermissionError:
